backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Redis and Celery
redis==5.0.1
celery==5.3.6

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0

# AI/OpenAI
openai==1.12.0
tiktoken==0.5.2

# Validation and Config
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==24.1.0
